+++ process/tests/pipeline.rs
@@ -1,17 +1,16 @@
-use process::{Cmd, Error};
+use process::{Command, Error};
 
 #[tokio::test]
-async fn pipeline() {
+async fn test_pipeline() {
     env_logger::builder().is_test(true).init();
 
-    let cmd = Cmd::from(vec!["echo hello", "cat"]);
+    let cmd = Command::from(vec!["echo hello", "cat"]);
     let out = cmd.run().await.unwrap().to_string_lossy();
-
     assert_eq!(out, "hello\n");
 
-    let cmd = Cmd::from(vec!["echo hello", "bad", "cat"]);
+    let cmd = Command::from(vec!["echo hello", "bad", "cat"]);
     match cmd.run().await.unwrap_err() {
-        Error::InvalidExitStatusCodeNonZeroError(cmd, status, err) => {
+        Error::GetExitStatusCodeNonZeroError(cmd, status, err) => {
             assert_eq!(cmd, "bad");
             assert_eq!(status, 127);
             assert_eq!(err, "sh: line 1: bad: command not found\n");